refactor(interfaces): split registry classes into separate files
- Extract HttpClientRegistry from http_interface.py to http_client_registry.py
- Extract WebSocketClientRegistry from websocket_interface.py to ws_client_registry.py
- Add name parameter to PipelineController
- Standardized logger naming to f"{self.__class__.__name__}_{self.name}"
