+++ src/interfaces/base/base_interface.py
@@ -16,7 +16,7 @@ class BaseInterface(Generic[TRegistry, TClient]):
     ) -> None:
         self.name: str = name if name else self.__class__.__name__
         self.client_registry: TRegistry = client_registry
-        self.logger = get_adapter(logger, self.name)
+        self.logger = get_adapter(logger, f"{self.__class__.__name__}_{self.name}")
         
         self.logger.info(f"{self.name} has been initialized.")
